rename code.js to code-legacy.js including test and snaps
