+++ botCommands/code-legacy.js
